Extract duplicated logic
