+++ azure/host.json
@@ -1,0 +1,3 @@
+{
+    "watchDirectories": [ "Shared" ]
+}
