+++ azure/test/formatter.test.js
@@ -1,41 +1,43 @@
 var assert = require('assert');
-var azure = require('../activity_logs_monitoring').forTests;
+var formatter = require('../shared/formatter.js').forTests;
+var constants = require('../shared/constants.js');
+
 
 describe('Azure Log Monitoring', function() {
-  describe('#getEventHubMessagesFormat', function() {
+  describe('#getLogFormat', function() {
     it('should return string', function() {
         eventHubMessages = '';
-        assert.equal(azure.STRING, azure.getEventHubMessagesFormat(eventHubMessages));
+        assert.equal(constants.STRING, formatter.getLogFormat(eventHubMessages));
         eventHubMessages = 'foobar';
-        assert.equal(azure.STRING, azure.getEventHubMessagesFormat(eventHubMessages));
+        assert.equal(constants.STRING, formatter.getLogFormat(eventHubMessages));
     });
     it('should return string array', function() {
       eventHubMessages = ['', 'foobar'];
-      assert.equal(azure.STRING_ARRAY, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.STRING_ARRAY, formatter.getLogFormat(eventHubMessages));
     });
     it('should return json with records', function() {
       eventHubMessages = [{'records': [{}, {}]}, {'records': [{}, {}]}];
-      assert.equal(azure.JSON_RECORDS, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.JSON_RECORDS, formatter.getLogFormat(eventHubMessages));
     });
     it('should return json object', function() {
       eventHubMessages = {'key': 'value', 'otherkey':'othervalue'};
-      assert.equal(azure.JSON_OBJECT, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.JSON_OBJECT, formatter.getLogFormat(eventHubMessages));
     });
     it('should return json no records', function() {
       eventHubMessages = [{'key': 'value', 'otherkey':'othervalue'}, {'key': 'value', 'otherkey':'othervalue'}];
-      assert.equal(azure.JSON_ARRAY, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.JSON_ARRAY, formatter.getLogFormat(eventHubMessages));
     });
     it('should return invalid', function() {
       eventHubMessages = 1;
-      assert.equal(azure.INVALID, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.INVALID, formatter.getLogFormat(eventHubMessages));
       eventHubMessages = () => {};
-      assert.equal(azure.INVALID, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.INVALID, formatter.getLogFormat(eventHubMessages));
       eventHubMessages = true;
-      assert.equal(azure.INVALID, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.INVALID, formatter.getLogFormat(eventHubMessages));
       eventHubMessages = null;
-      assert.equal(azure.INVALID, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.INVALID, formatter.getLogFormat(eventHubMessages));
       eventHubMessages = undefined;
-      assert.equal(azure.INVALID, azure.getEventHubMessagesFormat(eventHubMessages));
+      assert.equal(constants.INVALID, formatter.getLogFormat(eventHubMessages));
     });
   });
 
@@ -43,32 +45,32 @@ describe('Azure Log Monitoring', function() {
     it('should parse a valid record', function() {
       record = {'resourceId': '/SUBSCRIPTIONS/12345678-1234-ABCD-1234-1234567890AB/RESOURCEGROUPS/SOME-RESOURCE-GROUP/PROVIDERS/MICROSOFT.COMPUTE/VIRTUALMACHINES/SOME-VM'}
       expectedMetadata = {'tags': ["subscription_id:12345678-1234-abcd-1234-1234567890ab","resource_group:some-resource-group"], 'source': 'azure.compute'}
-      assert.deepEqual(expectedMetadata, azure.extractResourceId(record))
+      assert.deepEqual(expectedMetadata, formatter.extractResourceId(record))
     });
     it('should parse a valid record without provider', function() {
       record = {'resourceId': '/SUBSCRIPTIONS/12345678-1234-ABCD-1234-1234567890AB/RESOURCEGROUPS/SOME-RESOURCE-GROUP'}
       expectedMetadata = {'tags': ["subscription_id:12345678-1234-abcd-1234-1234567890ab","resource_group:some-resource-group"], 'source': ''}
-      assert.deepEqual(expectedMetadata, azure.extractResourceId(record))
+      assert.deepEqual(expectedMetadata, formatter.extractResourceId(record))
     });
     it('should parse a valid record without provider and resource group', function() {
       record = {'resourceId': '/SUBSCRIPTIONS/12345678-1234-ABCD-1234-1234567890AB'}
       expectedMetadata = {'tags': ["subscription_id:12345678-1234-abcd-1234-1234567890ab"], 'source': ''}
-      assert.deepEqual(expectedMetadata, azure.extractResourceId(record))
+      assert.deepEqual(expectedMetadata, formatter.extractResourceId(record))
     });
     it('should not fail on record without resourceId', function() {
       record = {'key':'value'}
       expectedMetadata = {'tags': [], 'source': ''}
-      assert.deepEqual(expectedMetadata, azure.extractResourceId(record))
+      assert.deepEqual(expectedMetadata, formatter.extractResourceId(record))
     });
     it('should not fail on string record', function() {
       record = {'key':'value'}
       expectedMetadata = {'tags': [], 'source': ''}
-      assert.deepEqual(expectedMetadata, azure.extractResourceId(record))
+      assert.deepEqual(expectedMetadata, formatter.extractResourceId(record))
     });
     it('should not fail on improper resourceId', function() {
       record = {'resourceId': 'foo/bar'}
       expectedMetadata = {'tags': [], 'source': ''}
-      assert.deepEqual(expectedMetadata, azure.extractResourceId(record))
+      assert.deepEqual(expectedMetadata, formatter.extractResourceId(record))
     });
   })
 });
